Time: 1075 ms (100.00%), Space: 79.1 MB (100.00%) - LeetHub

+++ 2306-naming-a-company/2306-naming-a-company.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long distinctNames(vector<string>&nums)
     {
-        int dp[26][26];
+        long long  dp[26][26];
         memset(dp,0,sizeof(dp));
         unordered_set<string>vis(nums.begin(),nums.end());
         for(auto &it:nums)
